reimplemented shoppingCartView to be compatible with propertychange, work in progress
